Manually copied changes Jake using shortcodes & templates
ITEM SHORTCODE TEMPLATES | This is a shortcode which can be pasted on the content area. Our goal here is to allow ourselves the flexibility of pulling information on-demand and also use CSS (or possibly unique templates) for the sake of creating variations of how the content will be displayed. The focus here is to allow us the flexibility to create multiple templates with ease.